refactor(backend): consolidate migration files and restructure host port entities
- Remove seed data generation command (cmd/seed/main.go)
- Consolidate database migrations into single init schema file
- Rename ip_address DTO to host_port for consistency
- Add host_port_snapshot DTO and model for snapshot tracking
- Rename host_port handler and repository files for clarity
- Implement host_port_snapshot service layer with CRUD operations
- Update website_snapshot service to work with new host_port structure
- Enhance terminal login UI with focus state tracking and Tab key navigation
- Update docker-compose configuration for development environment
- Refactor directory and website snapshot DTOs for improved data structure
- Add comprehensive test coverage for model and handler changes
- Simplify database schema by consolidating related migrations into single initialization file

+++ go-backend/internal/dto/directory.go
@@ -14,9 +14,9 @@ type DirectoryResponse struct {
 	TargetID      int       `json:"targetId"`
 	URL           string    `json:"url"`
 	Status        *int      `json:"status"`
-	ContentLength *int64    `json:"contentLength"`
+	ContentLength *int      `json:"contentLength"`
 	ContentType   string    `json:"contentType"`
-	Duration      *int64    `json:"duration"`
+	Duration      *int      `json:"duration"`
 	CreatedAt     time.Time `json:"createdAt"`
 }
 
@@ -34,9 +34,9 @@ type BulkCreateDirectoriesResponse struct {
 type DirectoryUpsertItem struct {
 	URL           string `json:"url" binding:"required,url"`
 	Status        *int   `json:"status"`
-	ContentLength *int64 `json:"contentLength"`
+	ContentLength *int   `json:"contentLength"`
 	ContentType   string `json:"contentType"`
-	Duration      *int64 `json:"duration"`
+	Duration      *int   `json:"duration"`
 }
 
 // BulkUpsertDirectoriesRequest represents bulk upsert directories request
