+++ go-backend/internal/model/directory.go
@@ -10,9 +10,9 @@ type Directory struct {
 	TargetID      int       `gorm:"column:target_id;not null;index:idx_directory_target;uniqueIndex:unique_directory_url_target,priority:1" json:"targetId"`
 	URL           string    `gorm:"column:url;size:2000;not null;index:idx_directory_url;uniqueIndex:unique_directory_url_target,priority:2" json:"url"`
 	Status        *int      `gorm:"column:status;index:idx_directory_status" json:"status"`
-	ContentLength *int64    `gorm:"column:content_length" json:"contentLength"`
+	ContentLength *int      `gorm:"column:content_length" json:"contentLength"`
 	ContentType   string    `gorm:"column:content_type;size:200" json:"contentType"`
-	Duration      *int64    `gorm:"column:duration" json:"duration"`
+	Duration      *int      `gorm:"column:duration" json:"duration"`
 	CreatedAt     time.Time `gorm:"column:created_at;autoCreateTime;index:idx_directory_created_at" json:"createdAt"`
 
 	// Relationships
